Fix wide characters around floating window's edge
Even when is_wide_char is true, it could be hidden under some floating
window. If the floating window's content comes just after the wide
character. Then right half of the wide character should be the floating
window's content, otherwise the floating window's content is out of
position in that line.

| あああああああああああ | // buffer content
| ああ +----------+ ああ |
| ABCDE| floating |RSTUV | // floating window over wide character
| ああ | floating | ああ |
      ^^ Here, there is wide character but it's hidden under the
         floating window's left edge. This wide character should be
         either of "not rendered" or "rendered with single width (border
         char overwraps)".

If simply rendered with double width:

| あああああああああああ | // buffer content
| ああ +----------+ ああ |
| ABCDE| floating |RSTUV | // floating window over wide character
| あああ| floating | ああ |
       ^^ the floating window become out of place

Similarly, even when is_wide_char is false, it could be the floating
window's rightmost character and the character on the buffer under this
floating window could be a left half of wide character. Then, wide
character hidden under the floating window is not rendered, but we need
to have a space right after the floating window's last content.

| あああああああああああ | // buffer content
| ああ +----------+ ああ |
| ABCDE| floating |RSTUV | // floating window over wide character
| ああ | floating | ああ |
                  ^^
      Here, there is single character (floating window's border symbol),
      but it should be rendered at double width since originally there
      was a double width character under the floating window.

If simply rendered with single width:

| あああああああああああ | // buffer content
| ああ +----------+ ああ |
| ABCDE| floating |RSTUV | // floating window over wide character
| ああ | floating |ああ |
                  ^^ content mistakenly packed to left